organic-chem-radar/requirements.txt
requests
feedparser
openai>=1.0.0
python-dotenv
